requirements.txt
pandas
google-generativeai
python-dotenv
Pillow
fpdf2
matplotlib
seaborn